Add helper to add background color on image
For some images with alpha channel, the background color and the
foreground color are the same after grey scale conversion, this will no
allow us to detect 1D or 2D code on the image.

The copy_image_on_background() allow to add a background color on a
image to avoid this issue.