qa: Replace flake8 by pylint
I am fine with keeping things that work, but Flake8 failed with the
following error:

    There was a critical error during execution of Flake8:
    plugin code for `flake8-builtins-unleashed[flake8_builtins]` does not match ^[A-Z]{1,3}[0-9]{0,3}$

I suppose that the `flake8-builtins-unleashed` plugin may not be
compatible with the current version of Flake8. Given that the plugin
was last released in 2017, I am not keen to debug further.

Actually, I tend to use pylint, not flake8. And I already have a known
good pylint configuration in TOML format, so here comes
"pyproject.toml"!